Make schema updates work on non-Windows platforms (#135)
* Make file paths separator-agnostic to fix non-Windows executions

Rework various constants in EiffelConstants to be java.nio.file.Path
objects instead of strings. That makes the intention of the constants
clear and we can easily work with Path objects to generate correct
paths on any platform.

* Make schema generation independent of order of files returned from OS

Previously all versions of the events were returned and they were all
converted and written to the same destination file. This happened to
work fine when the list of files returned from File.listFiles()
was in the correct order so that the latest version was the last file
to be processed. If that wasn't the case one or more schema files
didn't get the most recent version of the event.

We address this by using a version string parser to sort the filenames
in version order before picking the greatest version and generating a
schema from that file only.